refactor(component): 컴포넌트 의존성대로 폴더 구조 변경
