+++ src/components/main/Header.ts
@@ -1,4 +1,4 @@
-import Component from './Component';
+import Component from '../Component';
 
 interface HeaderProps {
   toggleModal: () => void;
