+++ src/components/main/Main.ts
@@ -1,8 +1,8 @@
-import { Category, SortBy, Restaurant } from '../type';
-import Component from './Component';
-import RestaurantList from './RestaurantList';
-import RestaurantFilterContainer from './RestaurantFilterContainer';
-import { DEFAULT_CATEGORY } from '../utils/constants';
+import { Category, SortBy, Restaurant } from '../../type';
+import Component from '../Component';
+import RestaurantFilterContainer from './restaurant/RestaurantFilterContainer';
+import RestaurantList from './restaurant/RestaurantList';
+import { DEFAULT_CATEGORY } from '../../utils/constants';
 
 interface MainProps {}
