+++ src/components/main/restaurant/RestaurantListItem.ts
@@ -1,6 +1,6 @@
-import { Restaurant } from '../type';
-import Component from './Component';
-import { CATEGORY_IMAGE_URL } from '../utils/constants';
+import { Restaurant } from '../../../type';
+import Component from '../../Component';
+import { CATEGORY_IMAGE_URL } from '../../../utils/constants';
 
 interface RestaurantListItemProps {
   restaurant: Restaurant;
